Disable OS unsupported C compiler variant keys
This is a hacky workaround that is needed because selectors do not work
with the new `compiler`s in `meta.yaml`. To fix this, we simply tell
`conda-build` to ignore the `c_compiler` key on Windows always and the
`m2w64_c_compiler` key on Unix always. This solves the issue when
selectors are unable to in a pretty similar way. Unfortunately we still
need the selectors in `build` as well (even though they don't work
correctly) as `conda-build` will complain that there are undefined
compilers (e.g. `m2w64_c`).