enable bootstrapping build in build.sh
if CONDA_BUILD_BOOTSTRAPPING is yes then the build is modified to remove dependencies
on other packages not yet built. a second round of building can create the package with
these build dependencies correctly specified.

Here CONDA_BUILD_BOOTSTRAPPING is set as a variable in a yaml file. This process should instead
be driven from the conda_build_config.yaml file.